Collapse juju application config per new spec
Refs: http://pad.lv/1621576
Refs: spec https://goo.gl/yqrrPI
