+++ cmd/juju/application/export_test.go
@@ -12,15 +12,8 @@ import (
 )
 
 // NewSetCommandForTest returns a SetCommand with the api provided as specified.
-func NewSetCommandForTest(serviceAPI serviceAPI) cmd.Command {
-	return modelcmd.Wrap(&setCommand{
-		serviceApi: serviceAPI,
-	})
-}
-
-// NewGetCommand returns a GetCommand with the api provided as specified.
-func NewGetCommandForTest(api getServiceAPI) cmd.Command {
-	return modelcmd.Wrap(&getCommand{
+func NewConfigCommandForTest(api configCommandAPI) cmd.Command {
+	return modelcmd.Wrap(&configCommand{
 		api: api,
 	})
 }
