+++ featuretests/cmdjuju_test.go
@@ -55,7 +55,7 @@ func (s *cmdJujuSuite) TestServiceSet(c *gc.C) {
 	ch := s.AddTestingCharm(c, "dummy")
 	svc := s.AddTestingService(c, "dummy-service", ch)
 
-	_, err := testing.RunCommand(c, application.NewSetCommand(), "dummy-service",
+	_, err := testing.RunCommand(c, application.NewConfigCommand(), "dummy-service",
 		"username=hello", "outlook=hello@world.tld")
 	c.Assert(err, jc.ErrorIsNil)
 
@@ -81,7 +81,7 @@ func (s *cmdJujuSuite) TestServiceUnset(c *gc.C) {
 	err := svc.UpdateConfigSettings(settings)
 	c.Assert(err, jc.ErrorIsNil)
 
-	_, err = testing.RunCommand(c, application.NewSetCommand(), "--to-default", "dummy-service", "username")
+	_, err = testing.RunCommand(c, application.NewConfigCommand(), "--reset", "dummy-service", "username")
 	c.Assert(err, jc.ErrorIsNil)
 
 	expect := charm.Settings{
@@ -118,7 +118,7 @@ settings:
 	ch := s.AddTestingCharm(c, "dummy")
 	s.AddTestingService(c, "dummy-service", ch)
 
-	context, err := testing.RunCommand(c, application.NewGetCommand(), "dummy-service")
+	context, err := testing.RunCommand(c, application.NewConfigCommand(), "dummy-service")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(testing.Stdout(context), jc.DeepEquals, expected)
 }
